@lemire 2022-04-10
I am skeptical that you can measure benefits from such a small micro-optimization without highly targeted benchmark, but the new code is clean and so there is no reason not to adopt it.