@Oppen 2022-04-10
I thought the same. My intention was to do a bit more work, but it ended up being more complex for little gain. I really saw a low hanging fruit in extracting the init out of the loop, but that actually made things slower for a test and made the code more complex due to still having to check outside that it wasn't in the right key, otherwise I'd iterate more than needed by resetting the container iterator.
I sent this part anyway because it seems almost free really.